% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_cleaning_functions.R
\name{find_duplicate_ids}
\alias{find_duplicate_ids}
\title{Find Duplicate IDs and Their Progress Values}
\usage{
find_duplicate_ids(myData, id = "Prolific_ID", progress = "Progress")
}
\arguments{
\item{myData}{A data frame containing the ID and progress columns.}

\item{id}{A character string specifying the column name of the IDs. Default is \code{"Prolific_ID"}.}

\item{progress}{A character string specifying the column name of the progress values. Default is \code{"Progress"}.}
}
\value{
A data frame containing duplicated IDs and their corresponding progress values.
}
\description{
This function identifies duplicate IDs in a dataset and retrieves the corresponding
progress values for each occurrence.
}
