% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_cleaning_functions.R
\name{remove_QP_default_fields}
\alias{remove_QP_default_fields}
\title{Remove irrelevant rows and columns from a Dataset}
\usage{
remove_QP_default_fields(myData)
}
\arguments{
\item{myData}{A data frame.}
}
\value{
A cleaned data frame with row 1 removed and specific columns dropped.
}
\description{
This function removes rows and columns introduced by Question Pro.
}
