% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_cleaning_functions.R
\name{remove_qualtrics_default_fields}
\alias{remove_qualtrics_default_fields}
\title{Remove irrelevant rows and columns from a dataset}
\usage{
remove_qualtrics_default_fields(myData)
}
\arguments{
\item{myData}{A data frame.}
}
\value{
A cleaned data frame with the first two rows removed and specific columns dropped.
}
\description{
This function removes the first two rows and specific columns that are irrelevant for analysis,
typically introduced by Qualtrics.
}
