% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_cleaning_functions.R
\name{remove_qualtrics_default_fields}
\alias{remove_qualtrics_default_fields}
\title{Remove irrelevant rows and columns from a Dataset}
\usage{
remove_qualtrics_default_fields(myData)
}
\arguments{
\item{myData}{A data frame.}
}
\value{
A data frame with rows 1-2 and columns 1-17 removed.
}
\description{
This function removes rows and columns introduced by Qualtrics.
}
